4.3. Operators(mergeMap)
16.mergeMap :-1:
we've only had one observable at a time. what happens if we're dealing with more than one observable? well, we have operators for that too. they're actually many, this was a bit confusing for me at first, so I'm going to keep it simple for now and just show you a few. we'll start with mergeMap