4.3. Operators(switchMap)
17.switchMap :-1:
but we don't want to fire off our observable for every single one ? well, then, we can use switchMap. Imagine you have a search bar. Every time your user types in there, you're going to query the results from a database. The imagine your user changes the search string while you're still waiting  for the query to return. switchMap is a great operator in that instance because it will cancel the old request and replace it with a new one automatically.