5.4. User Interface(Wrap an API)
20.Wrap an API :-1:
we're going to take input from that search bar and use it to search Reddit for cute picture of puppies and kittens.           async what's this ? this is the Async Pipe I mentioned earlier. This allows us to an observable straight to the DOM without using subscribe to unwrap it. The Async Pipe will automatically subscribe and the Async Pipe will automatically unsubscribe when the user leaves the page. But we should be pretty safe. okay , remember our search subject. rather than subscribing directly. we're going to create an observable called results. and we're going to hook that into our search subject. so we're going to put this .results here and now we don't need to subscribe because we're using the async pipe. we're going to keep that debounce Time and we're going to add a switch Map because just in case they start changing the search string while we're waiting for the query to come back we want it to cancel the old one and replace it. switch map will handle that for us automatically.              but I'm going to throw in a few bonus operators just because they're so cool.so I'm adding in distinct Unitl changed.
so  distinctUnitlChanged is just like it  sounds.  it's going to skip any values that are identical to the previous value. and .do() is a handy tittle thing that doesn't do anything. but it's very cool for console.log()s because now that we're not subscribing if we want to put a console.log() just for debugging we really don't want to subscribe just for that. so the .do() operator is a very handy thing that doesn't affect the observable at all